Separate test for subsets
