+++ bit/test_bit.py
@@ -115,6 +115,15 @@ class TestSuite(unittest.TestCase):
         random.shuffle(nums)
         self.assertEqual(single, single_number2(nums))
 
+    def test_subsets(self):
+
+        self.assertSetEqual(subsets([1, 2, 3]),
+                            {(), (1,), (2,), (3,), (1, 2), (1, 3), (2, 3), (1, 2, 3)})
+
+        self.assertSetEqual(subsets([10, 20, 30, 40]),
+                            {(10, 40), (10, 20, 40), (10, 30), (10, 20, 30, 40), (40,),
+                             (10, 30, 40), (30,), (20, 30), (30, 40), (10,), (),
+                             (10, 20), (20, 40), (20, 30, 40), (10, 20, 30), (20,)})
 
 if __name__ == '__main__':
     unittest.main()
